# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a API documentation for OB server.  You can find out more about Ocean Builders at [https://ocean.builders/](https://ocean.builders/)",
    "version": "0.7",
    "title": "Ocean Builder",
    "contact": {
      "email": "diego@scissorboy.com"
    }
  },
  "host": "oceanbuilders.herokuapp.com",
  "basePath": "/v1/api/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/lightining-scenes/{id}": {
      "post": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Create New LightScene",
        "description": "",
        "operationId": "createLightScene",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LightScene information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "source": {
                  "type": "string"
                },
                "sceneName": {
                  "type": "string"
                },
                "rooms": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "label": {
                        "type": "string"
                      },
                      "moodes": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "lightName": {
                              "type": "string"
                            },
                            "lightColor": {
                              "type": "string",
                              "pattern": "^0xFF[A-F0-9]{6}$"
                            },
                            "status": {
                              "type": "boolean"
                            },
                            "brightness": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 100
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "source": "seapod",
                "sceneName": "Day Mode",
                "rooms": [
                  {
                    "label": "BedRoom",
                    "moodes": [
                      {
                        "lightName": "Lightstrip 1",
                        "lightColor": "0xFF9B9591",
                        "status": false,
                        "brightness": 30
                      },
                      {
                        "lightName": "Lightstrip 2",
                        "lightColor": "0xFF123456",
                        "status": true,
                        "brightness": 50
                      }
                    ]
                  }
                ]
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "LightScene data is required, Invalid seapod id, LightScene validation faild, Duplicated Scene Name, LightScene source is not defiend"
          },
          "401": {
            "description": "Access denied. Not A member at the seapod"
          },
          "404": {
            "description": "Seapod Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update Lightscene Details",
        "description": "",
        "operationId": "updateLightScene",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LightScene information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "source": {
                  "type": "string"
                },
                "sceneName": {
                  "type": "string"
                },
                "rooms": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "label": {
                        "type": "string"
                      },
                      "moodes": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "lightName": {
                              "type": "string"
                            },
                            "lightColor": {
                              "type": "string",
                              "pattern": "^0xFF[A-F0-9]{6}$"
                            },
                            "status": {
                              "type": "boolean"
                            },
                            "brightness": {
                              "type": "integer",
                              "minimum": 0,
                              "maximum": 100
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "source": "seapod",
                "sceneName": "Day Mode",
                "rooms": [
                  {
                    "label": "BedRoom",
                    "moodes": [
                      {
                        "lightName": "Lightstrip 1",
                        "lightColor": "0xFF9B9591",
                        "status": false,
                        "brightness": 30
                      },
                      {
                        "lightName": "Lightstrip 2",
                        "lightColor": "0xFF123456",
                        "status": true,
                        "brightness": 50
                      }
                    ]
                  }
                ]
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "LightScene data is required, Invalid lightScene id, LightScene validation faild, Duplicated Scene Name, LightScene source is not defiend"
          },
          "401": {
            "description": "Access denied. Not A member at the seapod"
          },
          "404": {
            "description": "LightScene Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "delete": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Delete Lightscene",
        "description": "",
        "operationId": "deleteLightScene",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid lightScene id"
          },
          "404": {
            "description": "LightScene Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/lightining-scenes/{seapodId}/intensity/{value}": {
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update LightScene Intensity",
        "description": "",
        "operationId": "updateLightSceneIntensity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "path",
            "name": "value",
            "type": "integer",
            "required": true,
            "description": "value of intensity"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod id"
          },
          "401": {
            "description": "Access denied. Not A member at the seapod"
          },
          "404": {
            "description": "Seapod Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/lightining-scenes/{lightSceneId}/lightIntensity": {
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update Light Intensity",
        "description": "",
        "operationId": "updateLightIntensity",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Light information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "lightId": {
                  "type": "integer"
                },
                "intensity": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 100
                }
              },
              "example": {
                "lightId": "5f19a3f3a225d02fb8499557",
                "intensity": 80
              }
            }
          },
          {
            "in": "path",
            "name": "lightSceneId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid lightScene id"
          },
          "404": {
            "description": "LightScene Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/lightining-scenes/{seapodId}/status": {
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update LightScene Status",
        "description": "",
        "operationId": "updateLightSceneStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod id"
          },
          "401": {
            "description": "Access denied. Not A member at the seapod"
          },
          "404": {
            "description": "Seapod Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/lightining-scenes/{lightSceneId}/lightStatus": {
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update Light Status",
        "description": "",
        "operationId": "updateLightStatus",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Light information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "lightId": {
                  "type": "integer"
                }
              },
              "example": {
                "lightId": "5f19a3f3a225d02fb8499557"
              }
            }
          },
          {
            "in": "path",
            "name": "lightSceneId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid lightScene id"
          },
          "404": {
            "description": "LightScene Not Found, User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/lightining-scenes/{seapodId}/selected/{lightSceneId}": {
      "put": {
        "tags": [
          "Light Scenes"
        ],
        "summary": "Update Selected LightScene",
        "description": "",
        "operationId": "updateSelectedLightScene",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "path",
            "name": "lightSceneId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the lightScene"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid lightScene id, Invalid seapod id"
          },
          "401": {
            "description": "Access denied. Not A member at the seapod"
          },
          "404": {
            "description": "LightScene Not Found, User Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
  }
}